Implement NoFollowLinksBuilder

+++ Problem3/Checksums/Checksums/Directory.h
@@ -13,12 +13,12 @@ namespace Checksums
 		{
 		private:
 			vector<unique_ptr<FileTreeElement>> m_children;
-			size_t m_size;
+			uintmax_t m_size;
 		public:
 			Directory(string path);
 			void add(unique_ptr<FileTreeElement> child) override;
 			string getPath() const override;
-			size_t getSize() const override;
+			uintmax_t getSize() const override;
 			string toString() const override;
 		};
 	}
